+++ distribution-lifecycle/src/main/java/io/telicent/smart/cache/distribution/lifecycle/events/listeners/LoggingListener.java
@@ -16,12 +16,14 @@
 package io.telicent.smart.cache.distribution.lifecycle.events.listeners;
 
 import io.telicent.smart.cache.distribution.lifecycle.events.LifecycleAction;
+import lombok.ToString;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * A simple distribution lifecycle listener that logs lifecycle transitions
  */
+@ToString
 public class LoggingListener implements DistributionLifecycleListener {
     private static final Logger LOGGER = LoggerFactory.getLogger(LoggingListener.class);
 
@@ -30,4 +32,6 @@ public class LoggingListener implements DistributionLifecycleListener {
         LOGGER.info("Distribution {} transitioned from {} to {}", action.getDistributionId(),
                     action.getState().getFrom(), action.getState().getTo());
     }
+
+
 }